Rename files containing action functions
