+++ site/content/tutorial/12-actions/01-actions/app-a/click_outside.js
