+++ site/content/tutorial/12-actions/01-actions/app-b/click_outside.js
